tests: Move EditMessageTest class to test_message_edit.py.
